chore: update docs/moving functions from ModelUtils to Inflector
