+++ lib/adaptors/wallaby/custom/default_provider.rb
@@ -2,7 +2,7 @@
 
 module Wallaby
   class Custom
-    # Default authorization provider for {Wallaby::Custom} mode that whitelists everything.
+    # Default authorization provider for {Custom} mode that whitelists everything.
     class DefaultProvider < DefaultAuthorizationProvider
     end
   end
