+++ lib/adaptors/wallaby/custom/model_finder.rb
@@ -2,10 +2,10 @@
 
 module Wallaby
   class Custom
-    # Model finder for {Wallaby::Custom} mode that returns the list of model set by
+    # Model finder for {Custom} mode that returns the list of model set by
     # {Wallaby::Configuration#custom_models}
     class ModelFinder < ::Wallaby::ModelFinder
-      # @return [Wallaby::ClashArray] a list of classes
+      # @return [ClashArray] a list of classes
       def all
         Wallaby.configuration.custom_models
       end
