+++ lib/adaptors/wallaby/custom/model_service_provider.rb
@@ -2,7 +2,7 @@
 
 module Wallaby
   class Custom
-    # Model service provider for {Wallaby::Custom} mode
+    # Model service provider for {Custom} mode
     class ModelServiceProvider < ::Wallaby::ModelServiceProvider
     end
   end
