+++ lib/concerns/wallaby/decoratable.rb
@@ -5,11 +5,11 @@ module Wallaby
   module Decoratable
     # Get current model decorator. It comes from
     #
-    # - model decorator for {Wallaby::Configurable::ClassMethods#resource_decorator resource_decorator}
-    # - otherwise, model decorator for {Wallaby::Configurable::ClassMethods#application_decorator application_decorator}
+    # - model decorator for {Configurable::ClassMethods#resource_decorator resource_decorator}
+    # - otherwise, model decorator for {Configurable::ClassMethods#application_decorator application_decorator}
     #
     # Model decorator stores the information of **metadata** and **field_names** for **index**/**show**/**form** action.
-    # @return [Wallaby::ModelDecorator] current model decorator for this request
+    # @return [ModelDecorator] current model decorator for this request
     def current_model_decorator
       @current_model_decorator ||=
         current_decorator.try(:model_decorator) || \
@@ -18,9 +18,9 @@ module Wallaby
 
     # Get current resource decorator. It comes from
     #
-    # - {Wallaby::Configurable::ClassMethods#resource_decorator resource_decorator}
-    # - otherwise, {Wallaby::Configurable::ClassMethods#application_decorator application_decorator}
-    # @return [Wallaby::ResourceDecorator] current resource decorator for this request
+    # - {Configurable::ClassMethods#resource_decorator resource_decorator}
+    # - otherwise, {Configurable::ClassMethods#application_decorator application_decorator}
+    # @return [ResourceDecorator] current resource decorator for this request
     def current_decorator
       @current_decorator ||=
         DecoratorFinder.new(
@@ -41,7 +41,7 @@ module Wallaby
 
     # Wrap resource(s) with decorator(s).
     # @param resource [Object, Enumerable]
-    # @return [Wallaby::ResourceDecorator, Enumerable<Wallaby::ResourceDecorator>] decorator(s)
+    # @return [ResourceDecorator, Enumerable<Wallaby::ResourceDecorator>] decorator(s)
     def decorate(resource)
       return resource if resource.is_a?(ResourceDecorator)
       return resource.map { |item| decorate(item) } if resource.respond_to?(:map)
