+++ lib/concerns/wallaby/engineable.rb
@@ -6,11 +6,9 @@ module Wallaby
     # This helper method returns the current {Engine} routing proxy.
     # For example, if {Wallaby} is mounted at different paths at the same time:
     #
-    # ```
-    # mount Wallaby::Engine, at: '/admin'
-    # mount Wallaby::Engine, at: '/inner', as: :inner_engine,
-    #   defaults: { resources_controller: InnerController }
-    # ```
+    #     mount Wallaby::Engine, at: '/admin'
+    #     mount Wallaby::Engine, at: '/inner', as: :inner_engine,
+    #       defaults: { resources_controller: InnerController }
     #
     # If `/inner` is current request path, it returns `inner_engine` engine proxy.
     # @return [ActionDispatch::Routing::RoutesProxy] engine proxy for current request
@@ -36,8 +34,8 @@ module Wallaby
     end
 
     # @note This script name prefix is required for Rails
-    # {https://api.rubyonrails.org/classes/ActionView/RoutingUrlFor.html#method-i-url_for #url_for}
-    # to generate the correct URL.
+    #   {https://api.rubyonrails.org/classes/ActionView/RoutingUrlFor.html#method-i-url_for #url_for}
+    #   to generate the correct URL.
     # @return [String] current engine's script name
     def script_name
       current_engine_route.try { |route| route.path.spec.to_s }
