+++ lib/concerns/wallaby/paginatable.rb
@@ -3,11 +3,9 @@
 module Wallaby
   # Paginator related attributes
   module Paginatable
-    # Model paginator for current modal class. It comes from:
-    #
-    # - controller configuration {Wallaby::Configurable::ClassMethods#model_paginator .model_paginator}
-    # - a generic paginator based on {Wallaby::Configurable::ClassMethods#application_paginator .application_paginator}
+    # Model paginator for current modal class.
     # @return [Class] model paginator class
+    # @see PaginatorFinder#execute
     def current_paginator
       @current_paginator ||=
         PaginatorFinder.new(
@@ -26,8 +24,8 @@ module Wallaby
     # @param options [Hash]
     # @option options [Boolean] :paginate whether collection should be paginated
     # @return [#each]
-    # @see Wallaby::ModelServicer#paginate
-    def paginate(query, options)
+    # @see ModelServicer#paginate
+    def paginate(query, options = { paginate: true })
       options[:paginate] ? current_servicer.paginate(query, params) : query
     end
   end
