+++ lib/interfaces/wallaby/model_service_provider.rb
@@ -4,7 +4,7 @@ module Wallaby
   # Model service provider interface
   class ModelServiceProvider
     # @param model_class [Class]
-    # @param model_decorator [Wallaby::ModelDecorator, nil] model decorator
+    # @param model_decorator [ModelDecorator, nil] model decorator
     def initialize(model_class, model_decorator)
       raise ::ArgumentError, 'model class required' unless model_class
 
@@ -15,7 +15,7 @@ module Wallaby
     # To whitelist params for a model class
     # @param _params [ActionController::Parameters]
     # @param _action [String, Symbol]
-    # @param _authorizer [Wallaby::ModelAuthorizer]
+    # @param _authorizer [ModelAuthorizer]
     # @return [ActionController::Parameters] whitelisted params
     def permit(_params, _action, _authorizer)
       raise NotImplemented
