+++ lib/paginators/wallaby/model_paginator.rb
@@ -18,7 +18,7 @@ module Wallaby
     attr_reader :model_class
 
     # @!attribute [r] provider
-    # @return [Wallaby::ModelPaginationProvider] the instance that does the job
+    # @return [ModelPaginationProvider] the instance that does the job
     # @since wallaby-5.2.0
     attr_reader :provider
