--- lib/utils/wallaby/model_utils.rb
@@ -1,62 +1,0 @@
-# frozen_string_literal: true
-
-module Wallaby
-  # Model related utils
-  module ModelUtils
-    class << self
-      # Convert model class (e.g. `Namespace::Product`) into resources name (e.g. `namespace::products`)
-      # @param model_class [Class, String] model class
-      # @return [String] resources name
-      def to_resources_name(model_class)
-        return EMPTY_STRING if model_class.blank?
-
-        to_controllers_name(model_class).gsub SLASH, COLONS
-      end
-
-      # Convert model class (e.g. `Namespace::Product`) into resources name (e.g. `namespace/products`)
-      # @param model_class [Class, String] model class
-      # @return [String] controllers name
-      # @see https://github.com/rails/rails/blob/master/actionpack/lib/abstract_controller/base.rb#L135
-      def to_controllers_name(model_class)
-        return EMPTY_STRING if model_class.blank?
-
-        ActiveSupport::Inflector.tableize(model_class.to_s)
-      end
-
-      # Produce model label (e.g. `Namespace / Product`) for model class (e.g. `Namespace::Product`)
-      # @param model_class [Class, String] model class
-      # @return [String] model label
-      def to_model_label(model_class)
-        # TODO: change to use i18n translation
-        return EMPTY_STRING if model_class.blank?
-
-        model_class_name = to_model_name model_class
-        model_class_name.titleize.gsub(SLASH, SPACE + SLASH + SPACE)
-      end
-
-      # Convert resources name (e.g. `namespace::products`) into model class (e.g. `Namespace::Product`)
-      # @param resources_name [String] resources name
-      # @return [Class] model class
-      # @return [nil] when not found
-      def to_model_class(resources_name)
-        return if resources_name.blank?
-
-        class_name = to_model_name resources_name
-        # NOTE: DO NOT try to use const_defined? and const_get EVER.
-        # This is Rails, use constantize
-        class_name.constantize
-      rescue NameError
-        Logger.warn("Model #{class_name} cannot be found.", sourcing: 2..10)
-      end
-
-      # Convert resources name (e.g. `namespace::products`) into model name (e.g. `Namespace::Product`)
-      # @param resources_name [String] resources name
-      # @return [String] model name
-      def to_model_name(resources_name)
-        return EMPTY_STRING if resources_name.blank?
-
-        resources_name.to_s.singularize.gsub(COLONS, SLASH).camelize
-      end
-    end
-  end
-end
