+++ lib/wallaby/class_array.rb
@@ -20,7 +20,7 @@ module Wallaby
     # @!attribute [r] origin
     # @return [Array] The original array.
     def origin
-      # NOTE: DO NOT cache it by using instance variable!
+      # NOTE: DO NOT cache it using instance variable!
       @internal.map(&method(:to_class)).compact
     end
 
@@ -35,25 +35,25 @@ module Wallaby
     end
 
     # @param other [Array]
-    # @return [Wallaby::ClassArray] new Class array
+    # @return [ClassArray] new Class array
     def concat(other)
       self.class.new origin.concat(other.try(:origin) || other)
     end
 
     # @param other [Array]
-    # @return [Wallaby::ClassArray] new Class array
+    # @return [ClassArray] new Class array
     def -(other)
       self.class.new origin - (other.try(:origin) || other)
     end
 
     # @param item [Class, String]
-    # @return [Wallaby::ClassArray] self
+    # @return [ClassArray] self
     def <<(item)
       @internal << class_name_of(item)
       self
     end
 
-    # @return [Wallaby::ClassArray] self
+    # @return [ClassArray] self
     def each(&block)
       origin.each(&block)
       self
@@ -73,7 +73,7 @@ module Wallaby
     delegate :to_sentence, to: :origin
 
     # Ensure to freeze the {#internal}
-    # @return [Wallaby::ClassArray] self
+    # @return [ClassArray] self
     def freeze
       @internal.freeze
       super
