+++ lib/wallaby/core.rb
@@ -39,7 +39,6 @@ require 'utils/wallaby/inflector'
 require 'utils/wallaby/field_utils'
 require 'utils/wallaby/filter_utils'
 require 'utils/wallaby/locale'
-require 'utils/wallaby/model_utils'
 require 'utils/wallaby/module_utils'
 require 'utils/wallaby/params_utils'
 require 'utils/wallaby/utils'
