+++ spec/lib/wallaby/inflector_spec.rb
@@ -28,6 +28,31 @@ describe Wallaby::Inflector do
     end
   end
 
+  describe '.to_resources_name' do
+    it 'handles the namespace and returns resources name for a class name' do
+      expect(described_class.to_resources_name('Post')).to eq 'posts'
+      expect(described_class.to_resources_name('Wallaby::Post')).to eq 'wallaby::posts'
+      expect(described_class.to_resources_name('wallaby/posts')).to eq 'wallaby::posts'
+      expect(described_class.to_resources_name('Person')).to eq 'people'
+      expect(described_class.to_resources_name('Wallaby::Person')).to eq 'wallaby::people'
+      expect(described_class.to_resources_name('wallaby/person')).to eq 'wallaby::people'
+      expect(described_class.to_resources_name('Wallabies::Person')).to eq 'wallabies::people'
+      expect(described_class.to_resources_name('wallabies/person')).to eq 'wallabies::people'
+    end
+
+    context 'when model_class is blank' do
+      it 'returns blank string' do
+        expect(described_class.to_resources_name(nil)).to eq ''
+        expect(described_class.to_resources_name('')).to eq ''
+        expect(described_class.to_resources_name(' ')).to eq ''
+
+        expect(described_class.to_resources_name(false)).to eq ''
+        expect(described_class.to_resources_name([])).to eq ''
+        expect(described_class.to_resources_name({})).to eq ''
+      end
+    end
+  end
+
   describe '.to_controller_name' do
     it 'returns the controller class name' do
       expect(described_class.to_controller_name('/admin', 'categories')).to eq 'Admin::CategoriesController'
@@ -54,6 +79,58 @@ describe Wallaby::Inflector do
       expect(described_class.to_authorizer_name('/admin', 'category')).to eq 'Admin::CategoryAuthorizer'
       expect(described_class.to_authorizer_name('/admin', 'order::items')).to eq 'Admin::Order::ItemAuthorizer'
       expect(described_class.to_authorizer_name('/admin', 'order::item')).to eq 'Admin::Order::ItemAuthorizer'
+    end
+  end
+
+  describe '.to_model_label' do
+    it 'returns model label for a model class' do
+      expect(described_class.to_model_label('posts')).to eq 'Post'
+      expect(described_class.to_model_label('wallaby::posts')).to eq 'Wallaby / Post'
+      expect(described_class.to_model_label('post')).to eq 'Post'
+      expect(described_class.to_model_label('wallaby::post')).to eq 'Wallaby / Post'
+      expect(described_class.to_model_label('people')).to eq 'Person'
+      expect(described_class.to_model_label('wallaby::people')).to eq 'Wallaby / Person'
+      expect(described_class.to_model_label('wallabies::people')).to eq 'Wallabies / Person'
+
+      expect(described_class.to_model_label('person')).to eq 'Person'
+      expect(described_class.to_model_label('wallaby::person')).to eq 'Wallaby / Person'
+      expect(described_class.to_model_label('wallabies::person')).to eq 'Wallabies / Person'
+    end
+
+    context 'when model_class is blank' do
+      it 'returns blank string' do
+        expect(described_class.to_model_label(nil)).to eq ''
+        expect(described_class.to_model_label('')).to eq ''
+        expect(described_class.to_model_label(' ')).to eq ''
+
+        expect(described_class.to_model_label(false)).to eq ''
+        expect(described_class.to_model_label([])).to eq ''
+        expect(described_class.to_model_label({})).to eq ''
+      end
+    end
+  end
+
+  describe '.to_model_name' do
+    it 'returns model name for a resources name' do
+      expect(described_class.to_model_name('posts')).to eq 'Post'
+      expect(described_class.to_model_name('wallaby::posts')).to eq 'Wallaby::Post'
+      expect(described_class.to_model_name('post')).to eq 'Post'
+      expect(described_class.to_model_name('wallaby::post')).to eq 'Wallaby::Post'
+      expect(described_class.to_model_name('people')).to eq 'Person'
+      expect(described_class.to_model_name('wallaby::people')).to eq 'Wallaby::Person'
+      expect(described_class.to_model_name('wallabies::people')).to eq 'Wallabies::Person'
+    end
+
+    context 'when resources_name is blank' do
+      it 'returns blank string' do
+        expect(described_class.to_model_name(nil)).to eq ''
+        expect(described_class.to_model_name('')).to eq ''
+        expect(described_class.to_model_name(' ')).to eq ''
+
+        expect(described_class.to_model_name(false)).to eq ''
+        expect(described_class.to_model_name([])).to eq ''
+        expect(described_class.to_model_name({})).to eq ''
+      end
     end
   end
 end
